backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Pillow==10.0.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
pytz==2025.2
setuptools==82.0.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
gunicorn==20.1.0
